+++ libsrc4/nc4internal.c
@@ -24,6 +24,7 @@
 #include "ncdispatch.h" /* from libdispatch */
 #include "ncutf8.h"
 #include <stdarg.h>
+#include "ncrc.h"
 
 /** @internal Number of reserved attributes. These attributes are
  * hidden from the netcdf user, but exist in the implementation
@@ -52,11 +53,6 @@ static const NC_reservedatt NC_reserved[] = {
     {NC_ATT_NC3_STRICT_NAME, READONLYFLAG|MATERIALIZEDFLAG},		/*_nc3_strict*/
 };
 #define NRESERVED (sizeof(NC_reserved) / sizeof(NC_reservedatt))  /*|NC_reservedatt|*/
-
-/* These hold the file caching settings for the library. */
-size_t nc4_chunk_cache_size = CHUNK_CACHE_SIZE;            /**< Default chunk cache size. */
-size_t nc4_chunk_cache_nelems = CHUNK_CACHE_NELEMS;        /**< Default chunk cache number of elements. */
-float nc4_chunk_cache_preemption = CHUNK_CACHE_PREEMPTION; /**< Default chunk cache preemption. */
 
 static int NC4_move_in_NCList(NC* nc, int new_id);
 
@@ -737,6 +733,7 @@ int
 nc4_var_list_add2(NC_GRP_INFO_T *grp, const char *name, NC_VAR_INFO_T **var)
 {
     NC_VAR_INFO_T *new_var = NULL;
+    NCglobalstate* gs = NC_getglobalstate();
 
     /* Allocate storage for new variable. */
     if (!(new_var = calloc(1, sizeof(NC_VAR_INFO_T))))
@@ -745,9 +742,9 @@ nc4_var_list_add2(NC_GRP_INFO_T *grp, const char *name, NC_VAR_INFO_T **var)
     new_var->container = grp;
 
     /* These are the HDF5-1.8.4 defaults. */
-    new_var->chunk_cache_size = nc4_chunk_cache_size;
-    new_var->chunk_cache_nelems = nc4_chunk_cache_nelems;
-    new_var->chunk_cache_preemption = nc4_chunk_cache_preemption;
+    new_var->chunkcache.size = gs->chunkcache.size;
+    new_var->chunkcache.nelems = gs->chunkcache.nelems;
+    new_var->chunkcache.preemption = gs->chunkcache.preemption;
 
     /* Now fill in the values in the var info structure. */
     new_var->hdr.id = ncindexsize(grp->vars);
@@ -2093,3 +2090,141 @@ NC4_move_in_NCList(NC* nc, int new_id)
     return stat;
 }
 
+/**************************************************/
+/* NCglobal state management */
+
+static NCglobalstate* nc_globalstate = NULL;
+
+static int
+NC_createglobalstate(void)
+{
+    int stat = NC_NOERR;
+    const char* tmp = NULL;
+    
+    if(nc_globalstate == NULL) {
+        nc_globalstate = calloc(1,sizeof(NCglobalstate));
+    }
+    /* Initialize struct pointers */
+    nc_globalstate->rcinfo = (struct NCRCinfo*)calloc(1,sizeof(struct NCRCinfo));
+    if(nc_globalstate == NULL) return NC_ENOMEM;
+
+    /* Get environment variables */
+    if(getenv(NCRCENVIGNORE) != NULL)
+        nc_globalstate->rcinfo->ignore = 1;
+    tmp = getenv(NCRCENVRC);
+    if(tmp != NULL && strlen(tmp) > 0)
+        nc_globalstate->rcinfo->rcfile = strdup(tmp);
+    /* Initialize chunk cache defaults */
+    nc_globalstate->chunkcache.size = CHUNK_CACHE_SIZE;            /**< Default chunk cache size. */
+    nc_globalstate->chunkcache.nelems = CHUNK_CACHE_NELEMS;        /**< Default chunk cache number of elements. */
+    nc_globalstate->chunkcache.preemption = CHUNK_CACHE_PREEMPTION; /**< Default chunk cache preemption. */
+    
+    return stat;
+}
+
+/* Get global state */
+NCglobalstate*
+NC_getglobalstate(void)
+{
+    if(nc_globalstate == NULL)
+        NC_createglobalstate();
+    return nc_globalstate;
+}
+
+void
+NC_freeglobalstate(void)
+{
+    if(nc_globalstate != NULL) {
+        nullfree(nc_globalstate->tempdir);
+        nullfree(nc_globalstate->home);
+        nullfree(nc_globalstate->cwd);
+        NC_rcclear(nc_globalstate->rcinfo);
+	free(nc_globalstate->rcinfo);
+	free(nc_globalstate);
+	nc_globalstate = NULL;
+    }
+}
+/**************************************************/
+/* Specific property functions */
+
+/**
+Provide a function to store global data alignment
+information.
+Repeated calls to nc_set_alignment will overwrite any existing values.
+
+If defined, then for every file created or opened after the call to
+nc_set_alignment, and for every new variable added to the file, the
+most recently set threshold and alignment values will be applied
+to that variable.
+
+The nc_set_alignment function causes new data written to a
+netCDF-4 file to be aligned on disk to a specified block
+size. To be effective, alignment should be the system disk block
+size, or a multiple of it. This setting is effective with MPI
+I/O and other parallel systems.
+
+This is a trade-off of write speed versus file size. Alignment
+leaves holes between file objects. The default of no alignment
+writes file objects contiguously, without holes. Alignment has
+no impact on file readability.
+
+Alignment settings apply only indirectly, through the file open
+functions. Call nc_set_alignment first, then nc_create or
+nc_open for one or more files. Current alignment settings are
+locked in when each file is opened, then forgotten when the same
+file is closed. For illustration, it is possible to write
+different files at the same time with different alignments, by
+interleaving nc_set_alignment and nc_open calls.
+
+Alignment applies to all newly written low-level file objects at
+or above the threshold size, including chunks of variables,
+attributes, and internal infrastructure. Alignment is not locked
+in to a data variable. It can change between data chunks of the
+same variable, based on a file's history.
+
+Refer to H5Pset_alignment in HDF5 documentation for more
+specific details, interactions, and additional rules.
+
+@param threshold The minimum size to which alignment is applied.
+@param alignment The alignment value.
+
+@return ::NC_NOERR No error.
+@return ::NC_EINVAL Invalid input.
+@author Dennis Heimbigner
+@ingroup datasets
+*/
+int
+nc_set_alignment(int threshold, int alignment)
+{
+    NCglobalstate* gs = NC_getglobalstate();
+    gs->alignment.threshold = threshold;
+    gs->alignment.alignment = alignment;
+    gs->alignment.defined = 1;
+    return NC_NOERR;
+}
+
+/**
+Provide get function to retrieve global data alignment
+information.
+
+The nc_get_alignment function return the last values set by
+nc_set_alignment.  If nc_set_alignment has not been called, then
+it returns the value 0 for both threshold and alignment.
+
+@param thresholdp Return the current minimum size to which alignment is applied or zero.
+@param alignmentp Return the current alignment value or zero.
+
+@return ::NC_NOERR No error.
+@return ::NC_EINVAL Invalid input.
+@author Dennis Heimbigner
+@ingroup datasets
+*/
+
+int
+nc_get_alignment(int* thresholdp, int* alignmentp)
+{
+    NCglobalstate* gs = NC_getglobalstate();
+    if(thresholdp) *thresholdp = gs->alignment.threshold;
+    if(alignmentp) *alignmentp = gs->alignment.alignment;
+    return NC_NOERR;
+}