+++ libsrc4/nc4var.c
@@ -97,11 +97,11 @@ NC4_get_var_chunk_cache(int ncid, int varid, size_t *sizep,
 
     /* Give the user what they want. */
     if (sizep)
-        *sizep = var->chunk_cache_size;
+        *sizep = var->chunkcache.size;
     if (nelemsp)
-        *nelemsp = var->chunk_cache_nelems;
+        *nelemsp = var->chunkcache.nelems;
     if (preemptionp)
-        *preemptionp = var->chunk_cache_preemption;
+        *preemptionp = var->chunkcache.preemption;
 
     return NC_NOERR;
 }